Add files via upload
This image shows what the form looks like. You filter which team's overview you want to see by selecting which conference they are in from the Conference table on the left. The Conference selection then filters on the Teams table below, which displays the teams from whatever conference is selected. I made the Conference table by making a table with a ConferenceID and ConferenceName variables. This assigns a value to each conference with ConferenceID, and then ConferenceName is displayed in the table. I made the Teams with a teamID, conferenceID, and teamName variables. The teamID gives each team a unique code and ties to the main table to display all the team's information, conferenceID ties to the Conference table so this team will be under the right conference when that conference is selected in the Conference table, and teamName gets displayed. In the textbox controls, I would assign whatever variable to that textbox and I wrote a few queries in SQL to combine multiple variables (home wins and home losses, total wins and total losses, etc) using Concatenation to display a teams total record as "8-4" for example.